requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
motor>=3.3.0
python-dotenv>=1.0.0
requests>=2.31.0
